requirements.txt

peak
coverage
docutils
path.py
pygments
pyflakes
rope
pylint
pep8
pep362
pythoscope
sourcecodegen==0.6.13
git+git://github.com/mattvonrocketstein/multimethods-py.git#egg=multimethods
git+git://github.com/mattvonrocketstein/pygenie.git#egg=pygenie

